+++ CHANGELOG.md
@@ -39,6 +39,7 @@
  - Plugins are now imported in the inmanta_plugins package to allow importing submodules (#507)
  - Added event listener to Environment Service (#1996)
  - Autostarted agents can load a new value for the autostart_agent_map setting without agent restart (#1839)
+ - Added protected environment option (#1997)
 
 # v 2020.1 (2020-02-19) Changes in this release: